Experimental
The ULE scheduler can break the natural affinity of the interrupt
thread with the core receiving the interruption event. If a core
receives a hardware or software interrupt, the common case outcome
should be for the interrupt thread to be scheduled immediately
on the local core. With this particular issue in the ULE scheduler
it becomes likely that the local core will decide to schedule
an interrupt thread on a different core immediately.

Submitted by: Marc De La Gueronniere
MFC after: 1 week
Sponsored by: Verisign, Inc.